requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
psycopg2-binary>=2.9.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.21.0
